Started Working on Admin, added a movie addition options.
Now can add movies as admin account, need to add  option to check if the
user is an admin in order to see/use this functions.
Need to add Admin products management' and to return to Movie Page and
add all the function that is mentioned in the previous commit.