Finished the project for Now.
Next time thing to do:
Functions: finish the movie page, need to see it's genres and release
date, runtime.
Add the admin movie addition page somewhere
Maybe in the previous orders let admin view all orders in the site?

Testing: i have a lot to test just to bare in mind, maybe given
exceptions a popup message for the user?

Styling: of course i need to style the app to look more decent.

Cleaning: i need to organize the code and delete a lot of unnecessary,
make duplicate code as functions and more.

Documentation: when i will get to there i will check what to do.

That is all for now.